Module turbopack_dev_server::source::combined

Structs

CombinedContentSource

Combines multiple ContentSources by trying all content sources in order. The content source which responds with the most specific response (that is not a [ContentSourceContent::NotFound]) will be returned.

CombinedContentSourceVc

Vc for CombinedContentSource

PausableCombinedContentSource

A helper source which allows the CombinedContentSource to be paused while we ask for vary data.

PausableCombinedContentSourceVc

Vc for PausableCombinedContentSource

Type Definitions

CombinedContentSourceReadRef

see [turbo_tasks::ReadRef]

PausableCombinedContentSourceReadRef

see [turbo_tasks::ReadRef]