Trait turbopack_dev_server::source::ContentSource

pub trait ContentSource {
    fn get(
        &self,
        path: &str,
        data: Value<ContentSourceData>
    ) -> ContentSourceResultVc;
    fn get_children(&self) -> ContentSourcesVc;
}

A source of content that the dev server uses to respond to http requests.

Required Methods

fn get(&self, path: &str, data: Value<ContentSourceData>) -> ContentSourceResultVc

Gets content by path and request data from the source. data is empty by default and will only be filled when returning NeedData. This is useful as this method call will be cached based on it’s arguments, so we want to make the arguments contain as little information as possible to increase cache hit ratio.

fn get_children(&self) -> ContentSourcesVc

Gets any content sources wrapped in this content source.